Add event listeners for clear and new book buttons
Added an event listener to the "clearButton" to clear the library and
all associated data structures, as well as remove displayed books and
reset the favorite books list. An empty library card is displayed with
an "add" button to prompt the user to add new books when the library is
empty. Implemented an event listener for the "newBookButton" to display
the new book modal when the button is clicked, allowing users to add
new books to the library.