Implement modal close when clicking outside
This code should close the modal when a user clicks outside the modal,
it works properly in Chrome, but I have noticed some issues in Firefox,
therefore the code is commented out at a moment. Also attached a "load"
event listener to the window to ensure that the books in the library
are displayed when the page loads.